Move processUrl to util
